Table Component broken down to simpler components
