+++ react-states/editable-table/src/components/cellInput/cellInput.css
@@ -1,0 +1,6 @@
+.name-input {
+  width: 200px;
+  padding: 4px;
+  font-size: 16px;
+  outline: none;
+}
