+++ react-states/editable-table/src/components/table/table.css
@@ -33,13 +33,6 @@ th {
   font-weight: bold;
 }
 
-.name-input {
-  width: 200px;
-  padding: 4px;
-  font-size: 16px;
-  outline: none;
-}
-
 select {
   width: 200px;
   padding: 8px;
