+++ react-states/editable-table/src/components/table/table.js
@@ -1,82 +1,6 @@
 import React, { useState } from "react";
 import "./table.css";
-
-function CellInput({
-  columnName,
-  initialValue,
-  options = [],
-  onCellInputChange,
-  onChangeCellState,
-}) {
-  function handleKeyUp(e) {
-    if (e.key === "Enter") {
-      onChangeCellState();
-    }
-  }
-  if (columnName === "name" || columnName == "email") {
-    return (
-      <input
-        type="text"
-        className="name-input"
-        value={initialValue}
-        onChange={(e) => onCellInputChange(e.target.value)}
-        onKeyUp={handleKeyUp}
-        onBlur={onChangeCellState}
-      />
-    );
-  }
-
-  if (columnName === "status" || columnName === "role") {
-    return (
-      <select
-        value={initialValue}
-        onChange={(e) => onCellInputChange(e.target.value)}
-        onKeyUp={handleKeyUp}
-      >
-        {options.map((option) => (
-          <option value={option}> {option} </option>
-        ))}
-      </select>
-    );
-  }
-}
-
-function TableCell({
-  cellState,
-  cellValue,
-  rowIndex,
-  columnName,
-  options = [],
-  setTargetCell,
-  setUpdatedUsers,
-}) {
-  function handleDoubleClick() {
-    setTargetCell({ index: rowIndex, key: columnName });
-  }
-
-  function handleCellInputChange(value) {
-    setUpdatedUsers(rowIndex, columnName, value);
-  }
-
-  function onChangeCellState() {
-    setTargetCell({ index: -1, key: "" });
-  }
-
-  return (
-    <td className="ddd" onDoubleClick={handleDoubleClick}>
-      {cellState === "display" && cellValue}
-      {cellState === "input" && (
-        <CellInput
-          onCellInputChange={handleCellInputChange}
-          columnName={columnName}
-          initialValue={cellValue}
-          options={options}
-          onChangeCellState={onChangeCellState}
-        />
-      )}
-    </td>
-  );
-}
+import { TableCell } from "../tableCell/tableCell";
 
 export function EditableTable({ users }) {
   const columns = ["Id", "Name", "Email", "Role", "Status"];
